% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_related.R
\name{amr_positives}
\alias{amr_positives}
\title{Identify Antimicrobial Resistance Positive Plasmids from Blast Results}
\usage{
amr_positives(blast.results)
}
\arguments{
\item{blast.results}{Blast results loaded from read_blast or from Global Env}
}
\value{
Two column DF of plasmid names and genes present
}
\description{
This function loads the imported blast results, identifies which plasmids
carry AMR genes at highest identity. May have issues with multiple genes per
plasmid, currently optimized for identifying one of two genes
}
\examples{
\dontrun{
amr_positives(blastdata)
}
}

