% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_related.R
\name{amr_presence}
\alias{amr_presence}
\title{Adds the AMR_gene column to report}
\usage{
amr_presence(report, pos.samples)
}
\arguments{
\item{report}{Dataframe of results produced by \code{\link{subsampler}} or \code{\link{combine_results}}}
}
\value{
Report with AMR_genes added
}
\description{
Appends the results of amr_positives to the report in column AMR_gene, missing have "-" instead
}
\examples{
amr_presence(report)
}
\seealso{
\code{\link{subsampler}}, \code{\link{combine_results}}
}

