% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_related.R
\name{blast_parser}
\alias{blast_parser}
\title{Blast Results Parser Function}
\usage{
blast_parser(blast.results)
}
\arguments{
\item{blast.results}{Blast results loaded from read_blast}
}
\value{
Blast table with pID adjusted by ratio of hit length to query length (larger as denominator)
}
\description{
This function loads the imported blast results, extracts desired columns, Create new column of ratio between hit
length to query length - higher as denominator, adjusts pID by this ratio. Any AMR results are removed from the returned df.
}
\examples{
blast_parser(example_blast_results)
}

