% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blastdata}
\alias{blastdata}
\title{Example Table of Blast Results}
\format{Dataframe.}
\source{
Strains graciously provided by the authors of the following papers:
PMID: 26769926
PMID: 25545311
PMID: 26824951
}
\usage{
data(blastdata)
}
\description{
Example Table of Blast Results
}
\examples{
data(blastdata)
}
\references{
None Yet
(\href{http://www.ncbi.nlm.nih.gov/}{PubMed})
}
\keyword{datasets}

