% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_related.R
\name{combine_results}
\alias{combine_results}
\title{Combines SRST2 and Blast results into a single dataframe}
\usage{
combine_results(sr, br)
}
\arguments{
\item{sr}{SRST2 results loaded from read_srst2}

\item{br}{Blast results parsed by blast_parser}
}
\value{
Seven column dataframe of SRST2 results now including INC groups
}
\description{
Combines blast and SRST2 results, cuts to desired columns (Sample,
Plasmid, Inc_group, Coverage, Divergence, Length, Clusterid), matches
plasmids to BR and appends simplified INC names, all future
modifications are done to this dataframe
}
\examples{
\dontrun{
combine_results(example_srst2_results, example_blast_results)
}
}

