% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_related.R
\name{define_colours}
\alias{define_colours}
\title{Defining Colours Based on a Column of Data}
\usage{
define_colours(report, column)
}
\arguments{
\item{report}{Dataframe of results produced by \code{\link{subsampler}} or \code{\link{combine_results}}}

\item{column}{Specify a column by name}
}
\value{
Named vector of colours, names are factor levels of column supplied
}
\description{
\preformatted{
This function uses RColorBrewer to produce palettes based on the factor levels of the identified column in a report.
   }
}
\examples{
\dontrun{
define_colours(report, "AMR_gene")
}
}

