% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_aux.R
\name{main}
\alias{main}
\title{Run everything}
\usage{
main(blast.file, srst2.file, coverage.filter = NA, sureness.filter = NA,
  length.filter = NA, combine.inc = NA, plotly.user, plotly.api,
  post.plotly = NA, main.title = "Plasmid Profiles")
}
\arguments{
\item{blast.file}{System location of blast results (tsv)}

\item{srst2.file}{System location of srst2 results (tsv)}

\item{coverage.filter}{Filters results below percent read coverage specified (eg. 80)}

\item{sureness.filter}{Filters results below sureness specified (eg. 0.75)}

\item{length.filter}{Filters plasmid sequences shorter than length specified (eg. 10000)}

\item{combine.inc}{Flag to ombine incompatibility sub-groups into their main type (set to 1)}

\item{main.title}{A title for the figure}
}
\value{
Saves output files in working directory
}
\description{
Run all the interim functions to produce jpg, csv, and html outputs.
}
\examples{
main("data/blast_results.tsv", "data/srst2_results.tsv", cov.filter=NA, sureness.filter=0.75, len.filter=10000, main.title="Example Results")
}

