% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_aux.R
\name{minmax}
\alias{minmax}
\title{Minmax}
\usage{
minmax(df, maxcol, mincol)
}
\arguments{
\item{df}{Dataframe}

\item{maxcol}{Column to normalize from 0 to 1}

\item{mincol}{Column to normalize from 0 to -1}
}
\value{
Dataframe sorted by sum of maxcol and mincol
}
\description{
Takes two columns of numerical data,
normalizes it to ranges from 0 to 1 (0 to -1 for minimums),
sums them, arranges by sum, then returns the sorted dataframe
}
\examples{
\dontrun{
 minmax(df, maxcol, mincol)
 }
}

