% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_aux.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(x)
}
\arguments{
\item{x}{Vector of values}
}
\value{
Normalized vector of values
}
\description{
Normalizes a vector of values to a range of 0-1
x - min(x)) / (max(x) - min(x)
}
\examples{
\dontrun{
 normalize(x)
 }
}

