% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_related.R
\name{order_report}
\alias{order_report}
\title{Order the Report}
\usage{
order_report(report, anonymize = NA)
}
\arguments{
\item{report}{Dataframe of results produced by \code{\link{subsampler}} or \code{\link{combine_results}}}

\item{anonymize}{Flag to anything other than NA to replace plasmid and sample names with generic names}
}
\value{
Ordered report
}
\description{
Order the report first by sample order (tree), then by incompatibility group, then by sureness on each plasmid
}
\examples{
\dontrun{
order_report(report)
}
}
\seealso{
\code{\link{subsampler}}, \code{\link{combine_results}}
}

