% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_related.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Create GGPLOT Heatmap}
\usage{
plot_heatmap(report, len.highlight = NA)
}
\arguments{
\item{report}{Dataframe of results}

\item{len.highlight}{If anything but NA will highlight the largest plasmid hit per incompatibility group}
}
\value{
GGPLOT plotted heatmap
}
\description{
Using a ggplot2 tile geometry this function will create a heatmap of values in the report
coloured by incompatibility group, with alpha values from the sureness score. The order of
samples is determined by order_report and plasmids by incompatibility group and sureness score.
}
\examples{
\dontrun{
plot_heatmap(report)
}
}

