% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_related.R
\name{read_blast}
\alias{read_blast}
\title{Blast file import function}
\usage{
read_blast(br.file)
}
\arguments{
\item{br.file}{System location of the blast file, no default.}
}
\value{
Dataframe of blast data with correct column headers.
}
\description{
This function imports the 25 column blast file and adds column headers while designating the base filename.
}
\examples{
\dontrun{
read_blast("/data/blast_results.tsv")
}
}

