% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_related.R
\name{read_srst2}
\alias{read_srst2}
\title{SRST2 file import function}
\usage{
read_srst2(srst2.file)
}
\arguments{
\item{srst2.file}{System location of the srst2 file, no default.}
}
\value{
Dataframe of srst2 data with correct column headers.
}
\description{
This function imports the 14 column SRST2 file. Kind of superfluous
}
\examples{
\dontrun{
read_srst2("/data/srst2_results.tsv")
}
}

