% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_related.R
\name{zetner_score}
\alias{zetner_score}
\title{Adds the Zetner Score column to report}
\usage{
zetner_score(report)
}
\arguments{
\item{report}{Dataframe of results produced by \code{\link{subsampler}} or \code{\link{combine_results}}}
}
\value{
Report with zetner score added
}
\description{
Runs mimmax function on Coverage and Divergence, returns sum of normalized Coverage with negative normalized Divergence
a value which is then normalized from 0 to 1.
}
\examples{
zetner_score(report)
}
\seealso{
\code{\link{subsampler}}, \code{\link{combine_results}}
}

