% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_related.R
\name{create_plotly}
\alias{create_plotly}
\title{Create Plotly Object}
\usage{
create_plotly(report, user, api.key, post = NA,
  title = "Plasmid Profiles: Sureness, Incompatibility Group",
  len.highlight = NA)
}
\arguments{
\item{report}{Dataframe of results}

\item{user}{User ID for plotly web publishing}

\item{api.key}{API key for plotly web publishing}

\item{post}{Flag determines whether or not to post to plotly (default NA, no post)}

\item{title}{Title of heatmap}

\item{len.highlight}{If anything but NA will highlight the largest plasmid hit per incompatibility group}
}
\value{
plotly object
}
\description{
Combines the tree, heatmap, and titles to create final heatmap image.
}
\examples{
\dontrun{
create_plotly(report, title="Outbreak Plasmid Profiles")
}
}

