% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{report}
\alias{report}
\title{Example Complete Report}
\format{Dataframe.}
\source{
Strains graciously provided by the authors of the following papers:
Complete Genome and Plasmid Sequences of Three Canadian Isolates of
Salmonella enterica subsp. enterica Serovar Heidelberg from Human
and Food Sources. 2016 Labbé et al.
PMID: 26769926

Complete Sequence of Four Multidrug-Resistant MOBQ1 Plasmids Harboring
blaGES-5 Isolated from Escherichia coli and Serratia marcescens
Persisting in a Hospital in Canada. 2015 Boyd et al.
PMID: 25545311

Colistin-Nonsusceptible Pseudomonas aeruginosa Sequence Type 654 with
blaNDM-1 Arrives in North America. 2016 Mataseje et al.
PMID: 26824951
}
\usage{
data(report)
}
\description{
Example Complete Report
}
\examples{
data(report)
}
\references{
None Yet
(\href{http://www.ncbi.nlm.nih.gov/}{PubMed})
}
\keyword{datasets}

