% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_aux.R
\name{save_files}
\alias{save_files}
\title{Save Files}
\usage{
save_files(report, plot.png = NA, report.csv = NA, webpage = NA,
  title = "Plasmid Profiles")
}
\arguments{
\item{report}{Dataframe of results}

\item{plot.png}{Do you want to save a png? (Anything but NA)}

\item{report.csv}{Do you want to save a text report? (Anything but NA)}

\item{webpage}{Do you want to save an interactive heatmap as html? (Anything but NA)}

\item{title}{Enter a title for the plot}
}
\value{
Named vector of colours, names are factor levels of column supplied
}
\description{
Save various files: JPG, CSV, HTML depending on parameters
}
\examples{
\dontrun{
 save_files(report, plot.png=1, report.csv=1, webpage=NA)
}
}

