% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_aux.R
\name{save_files}
\alias{save_files}
\title{Save Files Produced}
\usage{
save_files(report, plot.jpg = NA, report.csv = NA, webpage = NA,
  title = "Plasmid Profiles")
}
\arguments{
\item{report}{Dataframe of results}

\item{plot.jpg}{Do you want to save a jpg? (Anything but NA)}

\item{report.csv}{Do you want to save a text report? (Anything but NA)}

\item{webpage}{Do you want to save an interactive heatmap as html? (Anything but NA)}
}
\value{
Named vector of colours, names are factor levels of column supplied
}
\description{
This function uses RColorBrewer to produce palettes based on the factor levels of the identified column in a report.
}
\examples{
save_files(report, plot.jpg=1, report.csv=1, webpage=NA)
}

