% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srst2data}
\alias{srst2data}
\title{Example Table of SRST2 Results}
\format{Dataframe.}
\source{
Here's where I got it from
}
\usage{
data(srst2data)
}
\description{
Example Table of SRST2 Results
}
\examples{
data(srst2data)
}
\references{
None Yet
(\href{http://www.ncbi.nlm.nih.gov/}{PubMed})
}
\keyword{datasets}

